user_service/requirements.txt
django==2.2
djangorestframework==3.9.2
psycopg2>=2.7,<3.0
